requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
boto3==1.20.41
botocore==1.23.41
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.1.4
coverage==7.2.7
cssselect==1.2.0
dateparser==0.7.6
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
django-heroku==0.0.0
djangorestframework==3.12.4
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.12.2
frozenlist==1.3.3
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
idna==3.3
jieba3k==0.35.1
jmespath==0.10.0
joblib==1.3.1
lxml==4.9.3
multidict==6.0.4
nltk==3.8.1
openai==0.27.8
packaging==21.3
Pillow==10.0.0
pycodestyle==2.8.0
pygooglenews==0.1.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.2.0
regex==2022.3.15
requests==2.31.0
requests-file==1.5.1
requests-ip-rotator==1.0.10
s3transfer==0.5.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tinysegmenter==0.3
tldextract==3.4.4
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.14.0
yarl==1.9.2
